+++ js/create-comment.js
@@ -1,0 +1,56 @@
+import { getRandomNumber, getRandomArrayElement } from './util.js';
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const USERS_NAMES = [
+  'Александр',
+  'Олег',
+  'Дима',
+  'Николай',
+  'Таисия',
+  'Ольга',
+  'Юлия',
+  'Марина',
+  'Павел',
+  'Данил',
+  'Сергей',
+  'Алексей',
+  'Владимир',
+  'Юрий',
+  'Валентина',
+  'Галина',
+  'Светлана',
+  'Аркадий',
+  'Егор',
+  'Роман',
+  'Ярослав',
+  'Кирилл',
+  'Динар',
+  'Екатерина',
+  'Ирина',
+];
+
+// Создание счетчика ID
+let commentId = 0;
+
+const getUniqCommentId = () => {
+  commentId++;
+  return commentId;
+};
+
+// Создание структуры комментария
+const createComments = () => ({
+  id: getUniqCommentId(),
+  avatar: `img/avatar-${  getRandomNumber(1, 6)  }.svg`,
+  message: getRandomArrayElement(MESSAGES),
+  name: getRandomArrayElement(USERS_NAMES),
+});
+
+export { createComments };
